RUMM-2207 Apply server date provider from config
